ureq

Struct BodyBuilder

pub struct BodyBuilder { /* private fields */ }

Builder for creating a response body.

This is useful for testing, or for Middleware that returns another body than the requested one.

Example

use ureq::Body;
use ureq::http::Response;

let body = Body::builder()
    .mime_type("text/plain")
    .charset("utf-8")
    .data("Hello world!");

let mut response = Response::builder()
    .status(200)
    .header("content-type", "text/plain; charset=utf-8")
    .body(body)?;

let text = response
    .body_mut()
    .read_to_string()?;

assert_eq!(text, "Hello world!");

Implementations

impl BodyBuilder

pub fn mime_type(self, mime_type: impl Into<String>) -> Self

Set the mime type of the body.

This does not set any HTTP headers. Affects Body decoding.

use ureq::Body;

let body = Body::builder()
    .mime_type("text/plain")
    .data("Hello world!");

pub fn charset(self, charset: impl Into<String>) -> Self

Set the mime type of the body

This does not set any HTTP headers. Affects Body decoding.

use ureq::Body;

let body = Body::builder()
    .mime_type("text/plain")
    .charset("utf-8")
    .data("Hello world!");

pub fn limit(self, l: u64) -> Self

Limit how much data is to be released from the body.

This does not set any HTTP headers. Affects Body decoding.

use ureq::Body;

let body = Body::builder()
    .mime_type("text/plain")
    .charset("utf-8")
    .limit(5)
    // This will be limited to "Hello"
    .data("Hello world!");

pub fn data(self, data: impl Into<Vec<u8>>) -> Body

Creates the body data turned into bytes.

Will automatically limit the body reader to the lenth of the data.

pub fn reader(self, data: impl Read + Send + Sync + 'static) -> Body

Creates a body from a streaming reader.

The reader can be limited by using .limit() or that the reader reaches the end.