ureq

Struct ProxyBuilder

pub struct ProxyBuilder { /* private fields */ }

Builder for configuring a proxy.

Obtained via Proxy::builder().

Implementations

impl ProxyBuilder

pub fn host(self, host: &str) -> Self

Set the proxy hostname

Defaults to localhost. Invalid hostnames surface in ProxyBuilder::build().

pub fn port(self, port: u16) -> Self

Set the proxy port

Defaults to whatever is default for the chosen ProxyProtocol.

pub fn username(self, v: &str) -> Self

Set the username

Defaults to none. Invalid usernames surface in ProxyBuilder::build().

pub fn password(self, v: &str) -> Self

Set the password

If you want to set only a password, no username, i.e. https://secret@foo.com, you need to set it as ProxyBuilder::username().

Defaults to none. Invalid passwords surface in ProxyBuilder::build().

pub fn resolve_target(self, do_resolve: bool) -> Self

Whether to resolve the target host locally before calling the proxy.

true - resolve target host locally before calling proxy.

false - let proxy resolve the host.

For SOCKS4, this defaults to true, for all other protocols false. I.e. in the “normal” case, we let the proxy itself resolve host names.

pub fn no_proxy(self, expr: &str) -> Self

Add a NO_PROXY expression to not route proxy through.

Correct expressions are:

example.com -> Literally match example.com, but not sub.example.com

.example.com -> Match sub.example.com and foo.sub.example.com, but not example.com.

*.example.com -> Exactly like .example.com

* -> Match everything

Silently ignores expressions that are not on the above form.

pub fn build(self) -> Result<Proxy, Error>

Construct the Proxy