ureq::unversioned

Module resolver

Name resolvers.

NOTE resolver does not (yet) follow semver.

NOTE: Resolver is deep configuration of ureq and is not required for regular use.

Name resolving is pluggable. The resolver’s duty is to take a URI and translate it to a socket address (IP + port). This is done as a separate step in regular ureq use. The hostname is looked up and provided to the Connector.

In some situations it might be desirable to not do this lookup, or to use another system than DNS for it.

Structs

ArrayVec

Simple impl of an array behaving like a vec.

DefaultResolver

Default resolver implementation.

Traits

Resolver

Trait for name resolvers.

Type Aliases

ResolvedSocketAddrs

Addresses as returned by the resolver.