ureq::unversioned::transport

Struct ConnectionDetails

pub struct ConnectionDetails<'a> {
    pub uri: &'a Uri,
    pub addrs: ResolvedSocketAddrs,
    pub config: &'a Config,
    pub request_level: bool,
    pub resolver: &'a dyn Resolver,
    pub now: Instant,
    pub timeout: NextTimeout,
    pub current_time: Arc<dyn Fn() -> Instant + Send + Sync + 'static>,
    pub run_connector: Arc<dyn Fn(&ConnectionDetails<'_>) -> Result<Box<dyn Transport>, Error> + Send + Sync>,
}

The parameters needed to create a Transport.

Fields
uri: &'a Uri

Full uri that is being requested.

In the case of CONNECT (HTTP) proxy, this is the URI of the proxy, and the actual URI is in the proxied field.

addrs: ResolvedSocketAddrs

The resolved IP address + port for the uri being requested. See Resolver.

For proxies, whetherh this holds real addresses depends on Proxy::resolve_target().

config: &'a Config

The configuration.

Agent or Request level.

request_level: bool

Whether the config is request level.

resolver: &'a dyn Resolver

The resolver configured on Agent.

Typically the IP address of the host in the uri is already resolved to the addr property. However there might be cases where additional DNS lookups need to be made in the connector itself, such as resolving a SOCKS proxy server.

now: Instant

Current time.

Time the ConnectionDetails was created.

timeout: NextTimeout

The next timeout for making the connection.

current_time: Arc<dyn Fn() -> Instant + Send + Sync + 'static>

Provides the current time.

run_connector: Arc<dyn Fn(&ConnectionDetails<'_>) -> Result<Box<dyn Transport>, Error> + Send + Sync>

Run the connector chain.

Used for CONNECT proxy to establish a connection to the proxy server itself.

Implementations

impl<'a> ConnectionDetails<'a>

pub fn needs_tls(&self) -> bool

Tell if the requested socket need TLS wrapping.